Event handlers have been implemented, Guava has been added and work on table view's backend started

+++ app/src/main/java/org/supermarkettycoon/TSaveFile.java
@@ -5,18 +5,18 @@ import java.util.ArrayList;
 import java.util.List;
 
 class TBoughtProducts extends TProduct {
-  public int quantity;
-  public double price;
-  public int buydate;
+    public int quantity;
+    public double price;
+    public int buydate;
 }
 
 public class TSaveFile {
-  public String username;
-  public int money = 1000;
-  public int day = 1;
-  public int power = 10;
-  public List<TLicense> licenses = new ArrayList<TLicense>();
-  public List<TUpgrade> upgrades = new ArrayList<TUpgrade>();
-  public List<TBoughtProducts> products = new ArrayList<TBoughtProducts>();
+    public String username;
+    public double money = 1000;
+    public int day = 1;
+    public int power = 10;
+    public List<TLicense> licenses = new ArrayList<TLicense>();
+    public List<TUpgrade> upgrades = new ArrayList<TUpgrade>();
+    public List<TBoughtProducts> products = new ArrayList<TBoughtProducts>();
 }
